+++ services.js
@@ -46,9 +46,9 @@ export const Services = class {
         1000 * 60, // every minute
         // 1000 * 1, // every second
         () => {
-          if (this.clock && this.clock.visible) {
-            this.clock.redraw();
-          }
+          this.extension.docks.forEach((d) => {
+            d._onClock();
+          });
         },
         -1
       ),
@@ -56,9 +56,9 @@ export const Services = class {
         'calendar',
         1000 * 60 * 15,
         () => {
-          if (this.calendar && this.calendar.visible) {
-            this.calendar.redraw();
-          }
+          this.extension.docks.forEach((d) => {
+            d._onCalendar();
+          });
         },
         -1
       ),
@@ -215,7 +215,7 @@ export const Services = class {
       }
     } catch (err) {
       // fail silently - don't crash
-      log(err);
+      console.log(err);
       this._disableNotifications++;
     }
 
@@ -336,24 +336,6 @@ export const Services = class {
     // added devices will subsequently be on mounted events
   }
 
-  redraw() {
-    let widgets = [this.clockCanvas, this.calendar];
-    widgets.forEach((w) => {
-      if (w) {
-        w.settings = {
-          dark_color: this.extension.drawing_dark_color,
-          light_color: this.extension.drawing_light_color,
-          accent_color: this.extension.drawing_accent_color,
-          dark_foreground: this.extension.drawing_dark_foreground,
-          light_foreground: this.extension.drawing_light_foreground,
-          secondary_color: this.extension.drawing_secondary_color,
-          clock_style: this.extension.clock_style,
-        };
-        w.redraw();
-      }
-    });
-  }
-
   updateIcon(item, settings) {
     if (!item) {
       return;
@@ -365,6 +347,8 @@ export const Services = class {
 
     let { scaleFactor, iconSize, dock } = settings;
 
+    // todo move dots and badges here?
+
     // the trash
     if (this.extension.trash_icon && icon.icon_name.startsWith('user-trash')) {
       let new_icon = this.trashFull ? 'user-trash-full' : 'user-trash';
@@ -373,19 +357,18 @@ export const Services = class {
       }
     }
 
-    let didCreate = false;
-
     // clock
     if (icon.icon_name == 'org.gnome.clocks') {
       if (this.extension.clock_icon) {
         let clock = item._clock;
         if (!clock) {
-          clock = new Clock(CANVAS_SIZE);
-          this.clock = clock;
+          clock = new Clock(CANVAS_SIZE, dock.extension._widgetStyle);
+          dock._clock = clock;
           item._clock = clock;
           item._appwell.first_child.add_child(clock);
         }
         if (clock) {
+          clock._icon = icon;
           clock.width = item._icon.width;
           clock.height = item._icon.height;
           clock.set_scale(item._icon.scaleX, item._icon.scaleY);
@@ -393,9 +376,12 @@ export const Services = class {
           clock.translationX = item._icon.translationX;
           clock.translationY = item._icon.translationY;
           clock.show();
+          item._icon.visible = !clock._hideIcon;
         }
       } else {
-        this.clock?.hide();
+        let clock = item._clock;
+        item._icon.visible = true;
+        clock?.hide();
       }
     }
 
@@ -404,8 +390,8 @@ export const Services = class {
       if (this.extension.calendar_icon) {
         let calender = item._calender;
         if (!calender) {
-          calender = new Calendar(CANVAS_SIZE);
-          this.calender = calender;
+          calender = new Calendar(CANVAS_SIZE, dock.extension._widgetStyle);
+          dock._calender = calender;
           item._calender = calender;
           item._appwell.first_child.add_child(calender);
         }
@@ -417,14 +403,13 @@ export const Services = class {
           calender.translationX = item._icon.translationX;
           calender.translationY = item._icon.translationY;
           calender.show();
+          item._icon.visible = !calender._hideIcon;
         }
       } else {
-        this.calender?.hide();
+        let calender = item._calender;
+        item._icon.visible = true;
+        calender?.hide();
       }
-    }
-
-    if (didCreate) {
-      this.redraw();
     }
   }
 };